MO-10871 Fix shiboken crashing during mocha build
During mochapro build the shiboken executable would
crash with a invalid memory exception.

The issue was caused by debug warning messages referencing
invalid CXFile pointers. Occurs in clangutils.cpp
getFileName as the code tries to display the source of the
warning. This string is generated in clangparser.cpp parse,
but the call to clang_disposeTranslationUnit disposes of
the string before the warnings are output.

To avoid the problem the -w flag has been added to the
clang flags.