MO-11135: specify arch when invoking Python during PySide build
Since Python is build as a universal binary, it can try to load the
wrong architecture of shiboken2.so resulting in build failure. We now
use the Mac's 'arch' command to choose the correct architecture of
Python to run, based on the CMAKE_OSX_ARCHITECTURES variable (which must
only contain a single value).