Always find the correct Qt modules (Core;Network;XmlPatterns) regardless of PATH content
If PATH contains the location of a Qt installation different to the version being built with, then find_file()
can return the wrong location for the required Qt modules. This results in the wrong Qt files being copied over.

This commit uses  get_target_property(..., LOCATION) to locate the correct Qt module.