Fixed invalid use of stale CXFile when reporting Diagnostic errors.
Any errors that occur in the call to clang::parse() cause Diagnostic to be added to clang::Builder.

Once clang::parse() returns, we must not use the CXFile SourceLocation.file value as it is may be stale.
Hence, we store the filename in SourceLocation.filename as a QString and this QString is used when reporting Diagnostics.